readme.dl
interface kindle fire with MIDI Keyboard application to Microsoft Music Maker Windows Store edition.  PC software will not acknowledge USB MIDI connection, or Wi-Fi. Have a wireless mouse and keyboard with a USB adapter for an interface I hope to "hi-jack" to allow the kindle and pc to interact in a route already established.